Working on SA c version.

+++ src/cpp/sw/sa_pipeline/stage6_sa.cpp
@@ -8,13 +8,13 @@ private:
 public:
     ST6_OUT new_output = {
             0, false, 0, 0,
-            {((int) 0), ((int) 0)},
-            {((int) 0), ((int) 0)}
+            {0, 0},
+            {0, 0}
     };
     ST6_OUT old_output = {
             0, false, 0, 0,
-            {((int) 0), ((int) 0)},
-            {((int) 0), ((int) 0)}
+            {0, 0},
+            {0, 0}
     };
 
     void compute(ST5_OUT st5_input) {
